+++ examples/lit-kitchen-sink/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  extends: ['../../.eslintrc.js'],
-  env: {
-    browser: true,
-  },
-  rules: {
-    'import/extensions': [0],
-  },
-};