+++ examples/official-storybook/preview.js
@@ -1,4 +1,4 @@
-import { document } from 'global';
+import global from 'global';
 import React, { Fragment, useEffect } from 'react';
 import isChromatic from 'chromatic/isChromatic';
 import {
@@ -10,10 +10,11 @@ import {
   styled,
   useTheme,
 } from '@storybook/theming';
-import { DocsPage } from '@storybook/addon-docs';
 import { Symbols } from '@storybook/components';
 
 import addHeadWarning from './head-warning';
+
+const { document } = global;
 
 if (process.env.NODE_ENV === 'development') {
   if (!process.env.DOTENV_DEVELOPMENT_DISPLAY_WARNING) {
@@ -157,7 +158,6 @@ export const parameters = {
   },
   docs: {
     theme: themes.light,
-    page: () => <DocsPage subtitleSlot={({ kind }) => `Subtitle: ${kind}`} />,
   },
   controls: {
     presetColors: [