+++ lib/cli/src/project_types.ts
@@ -28,7 +28,8 @@ export type SupportedFrameworks =
   | 'rax'
   | 'aurelia'
   | 'html'
-  | 'web-components';
+  | 'web-components'
+  | 'server';
 
 export enum ProjectType {
   UNDETECTED = 'UNDETECTED',
@@ -56,6 +57,7 @@ export enum ProjectType {
   SVELTE = 'SVELTE',
   RAX = 'RAX',
   AURELIA = 'AURELIA',
+  SERVER = 'SERVER',
 }
 
 export const SUPPORTED_FRAMEWORKS: SupportedFrameworks[] = [