+++ lib/components/src/Zoom/Zoom.tsx
@@ -1,6 +1,8 @@
-import { window as globalWindow } from 'global';
+import global from 'global';
 import { ZoomElement as Element } from './ZoomElement';
 import { ZoomIFrame as IFrame } from './ZoomIFrame';
+
+const { window: globalWindow } = global;
 
 export const browserSupportsCssZoom = (): boolean => {
   try {