+++ lib/components/src/controls/Object.tsx
@@ -1,4 +1,4 @@
-import { window as globalWindow } from 'global';
+import global from 'global';
 import cloneDeep from 'lodash/cloneDeep';
 import React, {
   ComponentProps,
@@ -18,6 +18,8 @@ import type { ControlProps, ObjectValue, ObjectConfig } from './types';
 import { Form } from '../form';
 import { Icons, IconsProps } from '../icon/icon';
 import { IconButton } from '../bar/button';
+
+const { window: globalWindow } = global;
 
 type JsonTreeProps = ComponentProps<typeof JsonTree>;