+++ lib/components/src/tooltip/WithTooltip.tsx
@@ -1,10 +1,12 @@
 import React, { FunctionComponent, ReactNode, useCallback, useState, useEffect } from 'react';
 import { styled } from '@storybook/theming';
-import { document } from 'global';
+import global from 'global';
 
 import TooltipTrigger from 'react-popper-tooltip';
 import { Modifier, Placement } from '@popperjs/core';
 import { Tooltip } from './Tooltip';
+
+const { document } = global;
 
 // A target that doesn't speak popper
 const TargetContainer = styled.div<{ mode: string }>`