+++ lib/core-client/src/preview/url.ts
@@ -1,10 +1,12 @@
-import { history, document } from 'global';
+import global from 'global';
 import qs from 'qs';
 import deprecate from 'util-deprecate';
 import { StoreSelectionSpecifier, StoreSelection } from '@storybook/client-api';
 import { StoryId, ViewMode } from '@storybook/addons';
 
 import { parseArgsParam } from './parseArgsParam';
+
+const { history, document } = global;
 
 export function pathToId(path: string) {
   const match = (path || '').match(/^\/story\/(.+)/);
@@ -58,14 +60,15 @@ const getFirstString = (v: ValueOf<qs.ParsedQs>): string | void => {
 
 const deprecatedLegacyQuery = deprecate(
   () => 0,
-  `URL formats with \`selectedKind\` and \`selectedName\` query parameters are deprecated. 
-Use \`id=$storyId\` instead. 
+  `URL formats with \`selectedKind\` and \`selectedName\` query parameters are deprecated.
+Use \`id=$storyId\` instead.
 See https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#new-url-structure`
 );
 
 export const getSelectionSpecifierFromPath: () => StoreSelectionSpecifier = () => {
   const query = qs.parse(document.location.search, { ignoreQueryPrefix: true });
   const args = typeof query.args === 'string' ? parseArgsParam(query.args) : undefined;
+  const globals = typeof query.globals === 'string' ? parseArgsParam(query.globals) : undefined;
 
   let viewMode = getFirstString(query.viewMode) as ViewMode;
   if (typeof viewMode !== 'string' || !viewMode.match(/docs|story/)) {
@@ -77,7 +80,7 @@ export const getSelectionSpecifierFromPath: () => StoreSelectionSpecifier = () =
   const storyId = path ? pathToId(path) : getFirstString(query.id);
 
   if (storyId) {
-    return { storySpecifier: storyId, args, viewMode, singleStory };
+    return { storySpecifier: storyId, args, globals, viewMode, singleStory };
   }
 
   // Legacy URL format
@@ -86,7 +89,7 @@ export const getSelectionSpecifierFromPath: () => StoreSelectionSpecifier = () =
 
   if (kind && name) {
     deprecatedLegacyQuery();
-    return { storySpecifier: { kind, name }, args, viewMode, singleStory };
+    return { storySpecifier: { kind, name }, args, globals, viewMode, singleStory };
   }
   return null;
 };