+++ scripts/build-manager.js
@@ -4,6 +4,12 @@ process.env.NODE_ENV = 'production';
 
 buildStaticStandalone({
   ignorePreview: true,
-  outputDir: './lib/core-server/prebuilt',
+  outputDir: './lib/manager-webpack4/prebuilt',
+  configDir: './scripts/build-manager-config',
+});
+
+buildStaticStandalone({
+  ignorePreview: true,
+  outputDir: './lib/manager-webpack5/prebuilt',
   configDir: './scripts/build-manager-config',
 });